# backend/app/api/routes/profiles.py
import base64
import binascii
import hashlib
import hmac
import json
from collections.abc import Callable, Sequence
from typing import Annotated
from uuid import UUID

from fastapi import APIRouter, Depends, Query
from sqlalchemy.orm import Session

from app.core.database import get_session
from app.core.errors import APIError
from app.db.models.profiles import CreatorContact, CreatorProfile, GameProfile
from app.repositories.profiles import CursorValue, ProfilesRepository
from app.schemas.common import CursorPage
from app.schemas.profiles import (
    CreatorContactResponse,
    CreatorManualUpdate,
    CreatorProfileCard,
    CreatorProfileDetail,
    FavoriteUpdate,
    GameProfileCard,
    GameProfileDetail,
    public_json_object,
)


PROFILE_TYPES = frozenset({"games", "creators"})


def _unknown_profile_type() -> APIError:
    return APIError(
        status_code=404,
        code="profile_type_unknown",
        message="The requested profile type is not supported.",
    )


def _require_profile_type(profile_type: str) -> str:
    if profile_type not in PROFILE_TYPES:
        raise _unknown_profile_type()
    return profile_type


def _profile_not_found() -> APIError:
    return APIError(
        status_code=404,
        code="profile_not_found",
        message="The requested profile was not found.",
    )


def _invalid_cursor() -> APIError:
    return APIError(
        status_code=400,
        code="profile_cursor_invalid",
        message="The profile cursor is invalid.",
    )


def _cursor_scope(
    profile_type: str, *, query: str, only_collection: bool
) -> dict[str, str | bool]:
    return {
        "profile_type": profile_type,
        "query": query.strip().casefold(),
        "only_collection": only_collection,
    }


def _cursor_signature(payload: dict[str, object], *, signing_key: bytes) -> str:
    canonical = json.dumps(
        payload,
        sort_keys=True,
        separators=(",", ":"),
        ensure_ascii=False,
    ).encode("utf-8")
    return hmac.new(signing_key, canonical, hashlib.sha256).hexdigest()


def _encode_cursor(
    sort_name: str,
    profile_id: UUID,
    *,
    profile_type: str,
    query: str,
    only_collection: bool,
    signing_key: bytes,
) -> str:
    signed_payload: dict[str, object] = {
        "v": 1,
        "key": [sort_name, str(profile_id)],
        "scope": _cursor_scope(
            profile_type, query=query, only_collection=only_collection
        ),
    }
    payload = {
        **signed_payload,
        "signature": _cursor_signature(signed_payload, signing_key=signing_key),
    }
    raw = json.dumps(
        payload, sort_keys=True, separators=(",", ":"), ensure_ascii=False
    ).encode("utf-8")
    return base64.urlsafe_b64encode(raw).decode("ascii").rstrip("=")


def _decode_cursor(
    cursor: str | None,
    *,
    profile_type: str,
    query: str,
    only_collection: bool,
    signing_key: bytes,
) -> CursorValue | None:
    if cursor is None:
        return None
    if not cursor or len(cursor) > 2048:
        raise _invalid_cursor()
    try:
        padded = cursor + "=" * (-len(cursor) % 4)
        raw = base64.b64decode(padded, altchars=b"-_", validate=True)
        value = json.loads(raw.decode("utf-8"))
        if (
            not isinstance(value, dict)
            or set(value) != {"v", "key", "scope", "signature"}
            or type(value["v"]) is not int
            or value["v"] != 1
            or not isinstance(value["key"], list)
            or len(value["key"]) != 2
            or not isinstance(value["key"][0], str)
            or not value["key"][0]
            or len(value["key"][0]) > 255
            or not isinstance(value["key"][1], str)
            or not isinstance(value["scope"], dict)
            or not isinstance(value["signature"], str)
        ):
            raise ValueError("invalid cursor shape")
        expected_scope = _cursor_scope(
            profile_type, query=query, only_collection=only_collection
        )
        if value["scope"] != expected_scope:
            raise ValueError("cursor scope mismatch")
        signed_payload = {
            "v": value["v"],
            "key": value["key"],
            "scope": value["scope"],
        }
        expected_signature = _cursor_signature(signed_payload, signing_key=signing_key)
        if not hmac.compare_digest(value["signature"], expected_signature):
            raise ValueError("cursor signature mismatch")
        profile_id = UUID(value["key"][1])
        if str(profile_id) != value["key"][1]:
            raise ValueError("non-canonical UUID")
    except (binascii.Error, UnicodeDecodeError, json.JSONDecodeError, ValueError):
        raise _invalid_cursor() from None
    return value["key"][0], profile_id


def _canonical_status_is_stale(value: object) -> bool:
    if isinstance(value, str):
        return value.casefold() == "stale"
    if isinstance(value, dict):
        return any(
            _canonical_status_is_stale(value.get(key))
            for key in ("status", "state", "freshness")
            if key in value
        )
    return False


def _creator_youtube_is_stale(source_status: object) -> bool:
    if not isinstance(source_status, dict):
        return False
    if any(
        _canonical_status_is_stale(source_status.get(key))
        for key in ("status", "state", "freshness")
        if key in source_status
    ):
        return True
    if _canonical_status_is_stale(source_status.get("youtube")):
        return True
    if any(
        _canonical_status_is_stale(source_status.get(key))
        for key in ("youtube_status", "youtube_state", "youtube_freshness")
        if key in source_status
    ):
        return True
    sources = source_status.get("sources")
    return isinstance(sources, dict) and _canonical_status_is_stale(
        sources.get("youtube")
    )


def _selected_contact(
    creator: CreatorProfile, *, manual_only: bool = False
) -> CreatorContactResponse | None:
    active = [contact for contact in creator.contacts if contact.is_active]
    manual = sorted(
        (contact for contact in active if contact.is_manual),
        key=lambda contact: (contact.created_at, str(contact.id)),
    )
    if manual:
        selected = manual[0]
        source = "manual"
    else:
        if manual_only:
            return None
        validation_rank = {
            "verified": 3,
            "valid": 2,
            "unverified": 1,
            "invalid": 0,
        }
        discovered = sorted(
            (contact for contact in active if not contact.is_manual),
            key=lambda contact: (
                -contact.priority,
                -validation_rank.get(contact.validation_state.casefold(), -1),
                contact.created_at,
                str(contact.id),
            ),
        )
        if not discovered:
            return None
        selected = discovered[0]
        source = selected.source_type
    return CreatorContactResponse(
        email=selected.email,
        source=source,
        source_url=selected.source_url,
        validation_state=selected.validation_state,
    )


def _game_card(profile: GameProfile) -> GameProfileCard:
    return GameProfileCard(
        id=profile.id,
        name=profile.sort_name,
        steam_app_id=profile.steam_app_id,
        canonical_url=profile.canonical_url,
        favorite=profile.favorite,
        current_facts=public_json_object(profile.current_facts),
        brief=public_json_object(profile.brief),
        source_status=public_json_object(profile.source_status),
        last_analyzed_at=profile.last_analyzed_at,
        next_analysis_at=profile.next_analysis_at,
    )


def _game_detail(profile: GameProfile) -> GameProfileDetail:
    return GameProfileDetail(
        **_game_card(profile).model_dump(),
        analysis=public_json_object(profile.analysis),
        model_metadata=public_json_object(profile.model_metadata),
        prompt_metadata=public_json_object(profile.prompt_metadata),
    )


def _creator_card(profile: CreatorProfile) -> CreatorProfileCard:
    stale = _creator_youtube_is_stale(profile.source_status)
    current_facts = {} if stale else profile.current_facts
    brief = {} if stale else profile.brief
    return CreatorProfileCard(
        id=profile.id,
        name=profile.sort_name,
        youtube_channel_id=profile.youtube_channel_id,
        canonical_url=profile.canonical_url,
        favorite=profile.favorite,
        current_facts=public_json_object(current_facts),
        brief=public_json_object(brief),
        source_status=public_json_object(profile.source_status),
        last_analyzed_at=profile.last_analyzed_at,
        next_analysis_at=profile.next_analysis_at,
        contact=_selected_contact(profile, manual_only=stale),
    )


def _creator_detail(profile: CreatorProfile) -> CreatorProfileDetail:
    analysis = (
        {} if _creator_youtube_is_stale(profile.source_status) else profile.analysis
    )
    return CreatorProfileDetail(
        **_creator_card(profile).model_dump(),
        analysis=public_json_object(analysis),
        model_metadata=public_json_object(profile.model_metadata),
        prompt_metadata=public_json_object(profile.prompt_metadata),
        manual_notes=profile.manual_notes,
    )


def _next_cursor(
    rows: Sequence[GameProfile] | Sequence[CreatorProfile],
    *,
    has_more: bool,
    profile_type: str,
    query: str,
    only_collection: bool,
    signing_key: bytes,
) -> str | None:
    if not has_more or not rows:
        return None
    last = rows[-1]
    return _encode_cursor(
        last.sort_name,
        last.id,
        profile_type=profile_type,
        query=query,
        only_collection=only_collection,
        signing_key=signing_key,
    )


def _validated_cursor(
    repository: ProfilesRepository,
    profile_type: str,
    raw_cursor: str | None,
    *,
    query: str,
    only_collection: bool,
    signing_key: bytes,
) -> CursorValue | None:
    cursor = _decode_cursor(
        raw_cursor,
        profile_type=profile_type,
        query=query,
        only_collection=only_collection,
        signing_key=signing_key,
    )
    if cursor is not None and not repository.cursor_matches(
        profile_type,
        cursor,
        query=query,
        only_collection=only_collection,
    ):
        raise _invalid_cursor()
    return cursor


def create_router(
    authenticate_workspace: Callable, *, cursor_signing_secret: str
) -> APIRouter:
    cursor_signing_key = hashlib.sha256(
        b"find-me-gamer/profile-cursor/v1\0" + cursor_signing_secret.encode("utf-8")
    ).digest()
    router = APIRouter(
        prefix="/api/v1/profiles",
        tags=["profiles"],
        dependencies=[Depends(authenticate_workspace)],
    )

    @router.get(
        "/games",
        response_model=CursorPage[GameProfileCard],
        operation_id="listGameProfiles",
    )
    def list_games(
        query: Annotated[str, Query(max_length=255)] = "",
        only_collection: bool = False,
        cursor: str | None = None,
        limit: Annotated[int, Query(ge=1, le=100)] = 50,
        database_session: Session = Depends(get_session),
    ) -> CursorPage[GameProfileCard]:
        repository = ProfilesRepository(database_session)
        decoded_cursor = _validated_cursor(
            repository,
            "games",
            cursor,
            query=query,
            only_collection=only_collection,
            signing_key=cursor_signing_key,
        )
        rows, has_more = repository.list_games(
            query=query,
            only_collection=only_collection,
            cursor=decoded_cursor,
            limit=limit,
        )
        return CursorPage(
            items=[_game_card(profile) for profile in rows],
            next_cursor=_next_cursor(
                rows,
                has_more=has_more,
                profile_type="games",
                query=query,
                only_collection=only_collection,
                signing_key=cursor_signing_key,
            ),
        )

    @router.get(
        "/creators",
        response_model=CursorPage[CreatorProfileCard],
        operation_id="listCreatorProfiles",
    )
    def list_creators(
        query: Annotated[str, Query(max_length=255)] = "",
        only_collection: bool = False,
        cursor: str | None = None,
        limit: Annotated[int, Query(ge=1, le=100)] = 50,
        database_session: Session = Depends(get_session),
    ) -> CursorPage[CreatorProfileCard]:
        repository = ProfilesRepository(database_session)
        decoded_cursor = _validated_cursor(
            repository,
            "creators",
            cursor,
            query=query,
            only_collection=only_collection,
            signing_key=cursor_signing_key,
        )
        rows, has_more = repository.list_creators(
            query=query,
            only_collection=only_collection,
            cursor=decoded_cursor,
            limit=limit,
        )
        return CursorPage(
            items=[_creator_card(profile) for profile in rows],
            next_cursor=_next_cursor(
                rows,
                has_more=has_more,
                profile_type="creators",
                query=query,
                only_collection=only_collection,
                signing_key=cursor_signing_key,
            ),
        )

    @router.patch(
        "/creators/{profile_id}/manual",
        response_model=CreatorProfileDetail,
        operation_id="updateCreatorManual",
    )
    def update_creator_manual(
        profile_id: UUID,
        update: CreatorManualUpdate,
        database_session: Session = Depends(get_session),
    ) -> CreatorProfileDetail:
        creator = ProfilesRepository(database_session).update_creator_manual(
            profile_id,
            contact_email=str(update.contact_email) if update.contact_email else None,
            notes=update.notes,
        )
        if creator is None:
            raise _profile_not_found()
        database_session.commit()
        return _creator_detail(creator)

    @router.get(
        "/{profile_type}/{profile_id}",
        response_model=GameProfileDetail | CreatorProfileDetail,
        operation_id="getProfile",
    )
    def read_profile(
        profile_type: Annotated[str, Depends(_require_profile_type)],
        profile_id: UUID,
        database_session: Session = Depends(get_session),
    ) -> GameProfileDetail | CreatorProfileDetail:
        repository = ProfilesRepository(database_session)
        if profile_type == "games":
            game = repository.get_game(profile_id)
            if game is None:
                raise _profile_not_found()
            return _game_detail(game)
        if profile_type == "creators":
            creator = repository.get_creator(profile_id)
            if creator is None:
                raise _profile_not_found()
            return _creator_detail(creator)
        raise _unknown_profile_type()

    @router.patch(
        "/{profile_type}/{profile_id}/favorite",
        response_model=GameProfileCard | CreatorProfileCard,
        operation_id="setProfileFavorite",
    )
    def set_favorite(
        profile_type: Annotated[str, Depends(_require_profile_type)],
        profile_id: UUID,
        update: FavoriteUpdate,
        database_session: Session = Depends(get_session),
    ) -> GameProfileCard | CreatorProfileCard:
        repository = ProfilesRepository(database_session)
        if profile_type == "games":
            game = repository.set_game_favorite(profile_id, favorite=update.favorite)
            if game is None:
                raise _profile_not_found()
            database_session.commit()
            return _game_card(game)
        if profile_type == "creators":
            creator = repository.set_creator_favorite(
                profile_id, favorite=update.favorite
            )
            if creator is None:
                raise _profile_not_found()
            database_session.commit()
            return _creator_card(creator)
        raise _unknown_profile_type()

    @router.get("/{profile_type}", operation_id="rejectUnknownProfileType")
    def reject_unknown_list_type(profile_type: str) -> None:
        raise _unknown_profile_type()

    return router