# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreatorManualUpdate": {
        "properties": {
          "contact_email": {
            "anyOf": [
              {
                "format": "email",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contact Email"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 20000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "required": [
          "contact_email",
          "notes"
        ],
        "title": "CreatorManualUpdate",
        "type": "object"
      },
...
      "CreatorProfileDetail": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "brief": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "canonical_url": {
            "title": "Canonical Url",
            "type": "string"
          },
          "contact": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CreatorContactResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_facts": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "favorite": {
            "title": "Favorite",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_analyzed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Analyzed At"
          },
          "manual_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Manual Notes"
          },
          "model_metadata": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "next_analysis_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Analysis At"
          },
          "prompt_metadata": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "source_status": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "type": {
            "const": "creator",
            "default": "creator",
            "title": "Type",
            "type": "string"
          },
          "youtube_channel_id": {
            "title": "Youtube Channel Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "youtube_channel_id",
          "canonical_url",
          "favorite",
          "current_facts",
          "brief",
          "source_status",
          "last_analyzed_at",
          "next_analysis_at",
          "contact",
          "analysis",
          "model_metadata",
          "prompt_metadata",
          "manual_notes"
        ],
        "title": "CreatorProfileDetail",
        "type": "object"
      },
      "CursorPage_CreatorProfileCard_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CreatorProfileCard"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CursorPage[CreatorProfileCard]",
        "type": "object"
      },
      "CursorPage_GameProfileCard_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/GameProfileCard"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CursorPage[GameProfileCard]",
        "type": "object"
      },
...
      "FavoriteUpdate": {
        "properties": {
          "favorite": {
            "title": "Favorite",
            "type": "boolean"
          }
        },
        "required": [
          "favorite"
        ],
        "title": "FavoriteUpdate",
        "type": "object"
      },
...
      "GameProfileDetail": {
        "properties": {
          "analysis": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "brief": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "canonical_url": {
            "title": "Canonical Url",
            "type": "string"
          },
          "current_facts": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "favorite": {
            "title": "Favorite",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "last_analyzed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Analyzed At"
          },
          "model_metadata": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "next_analysis_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Analysis At"
          },
          "prompt_metadata": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "source_status": {
            "$ref": "#/components/schemas/PublicJSONObject"
          },
          "steam_app_id": {
            "title": "Steam App Id",
            "type": "string"
          },
          "type": {
            "const": "game",
            "default": "game",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "steam_app_id",
          "canonical_url",
          "favorite",
          "current_facts",
          "brief",
          "source_status",
          "last_analyzed_at",
          "next_analysis_at",
          "analysis",
          "model_metadata",
          "prompt_metadata"
        ],
        "title": "GameProfileDetail",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PublicJSONObject": {
        "additionalProperties": {
          "$ref": "#/components/schemas/PublicJSONValue"
        },
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Find Me Gamer API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/profiles/creators": {
      "get": {
        "operationId": "listCreatorProfiles",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 255,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "only_collection",
            "required": false,
            "schema": {
              "default": false,
              "title": "Only Collection",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_CreatorProfileCard_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Creators",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/creators/{profile_id}/manual": {
      "patch": {
        "operationId": "updateCreatorManual",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatorManualUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatorProfileDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Creator Manual",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/games": {
      "get": {
        "operationId": "listGameProfiles",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": false,
            "schema": {
              "default": "",
              "maxLength": 255,
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "only_collection",
            "required": false,
            "schema": {
              "default": false,
              "title": "Only Collection",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CursorPage_GameProfileCard_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Games",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/{profile_type}": {
      "get": {
        "operationId": "rejectUnknownProfileType",
        "parameters": [
          {
            "in": "path",
            "name": "profile_type",
            "required": true,
            "schema": {
              "title": "Profile Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Unknown List Type",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/{profile_type}/{profile_id}": {
      "get": {
        "operationId": "getProfile",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "profile_type",
            "required": true,
            "schema": {
              "title": "Profile Type",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/GameProfileDetail"
                    },
                    {
                      "$ref": "#/components/schemas/CreatorProfileDetail"
                    }
                  ],
                  "title": "Response Getprofile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Profile",
        "tags": [
          "profiles"
        ]
      }
    },
    "/api/v1/profiles/{profile_type}/{profile_id}/favorite": {
      "patch": {
        "operationId": "setProfileFavorite",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Profile Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "profile_type",
            "required": true,
            "schema": {
              "title": "Profile Type",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FavoriteUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/GameProfileCard"
                    },
                    {
                      "$ref": "#/components/schemas/CreatorProfileCard"
                    }
                  ],
                  "title": "Response Setprofilefavorite"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Favorite",
        "tags": [
          "profiles"
        ]
      }
    },
...
  }
}